Update solution for tanh-activation

+++ tanh-activation/tanh-activation.py
@@ -6,4 +6,4 @@ def tanh(x):
     """
     # Write code here
     x = np.asarray(x, dtype=float)
-    return (np.exp(x) - np.exp(-x)) / (np.exp(x) + np.exp(-x))
+    return (np.exp(x) - np.exp(-x)) / (np.exp(x) + np.exp(-x)).astype(float)
